third_party/blink/web_tests/platform/generic/fast/css/add-remove-stylesheets-at-once-minimal-recalc-style-expected.txt

PASS internals.updateStyleAndReturnAffectedElementCount() is 8
PASS internals.updateStyleAndReturnAffectedElementCount() is 10
PASS getComputedStyle(document.querySelector('.foo')).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelector('.baz')).color is "rgb(0, 0, 255)"
PASS getComputedStyle(document.querySelector('.bar')).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE

foo
foo
foo
bar
bar
bar
baz
baz
baz
